requirements.txt
#For the tray icon
pystray==0.19.5
pillow==11.2.1